refactor(speech): extract state_machine test mocks+tests to sibling file
`voice_loop/state_machine.rs` was 2202 lines, dominated by a single
~1418-line `#[cfg(test)] mod mocks { … }` block holding the mock STT/TTS
backends and every `#[tokio::test]` for the loop. Move that block verbatim
into `voice_loop/state_machine/mocks.rs`, declared in `state_machine.rs` as
`#[cfg(test)] mod mocks;`.

Pure relocation, no behaviour change: the module stays
`voice_loop::state_machine::mocks`, so `super::` still resolves to the
`state_machine` module and every test reaches the loop's public types
unchanged. `state_machine.rs` drops 2202 → 786 lines (the actual loop);
the existing tests are the regression guard.

Follows the established `voice_loop/` "one concern per file" pattern and the
prior #266 extraction of the pure quit-phrase / TTS-markdown helpers.

Verification (all green):
- cargo test -p primer-speech --features voice-loop --lib: 101 passed,
  2 ignored — identical to baseline; all under
  voice_loop::state_machine::mocks::*
- cargo test --workspace: 0 failures
- cargo clippy --workspace --all-targets -- -D warnings: clean
- cargo clippy -p primer-speech --features voice-loop --all-targets: clean
- cargo fmt --all -- --check: clean
